requirements.txt
pymongo[srv]
Flask-API
python-dotenv
gradio
fpdf2